% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_packages.R
\name{check.packages}
\alias{check.packages}
\title{Check and reports if any of your packages are out of date with CRAN}
\usage{
check.packages()
}
\description{
Check and reports if any of your packages are out of date with CRAN
}

