% Generated by roxygen2 (4.0.0): do not edit by hand
\name{lsp}
\alias{lsp}
\title{List functions inside any package}
\usage{
lsp(package, pattern, all.names = FALSE)
}
\arguments{
\item{package}{Name of a loaded package}

\item{all.names}{Default is \code{FALSE}}

\item{pattern}{Filter function search by package}
}
\description{
List functions inside any package
}
\examples{
\dontrun{
library(ggplot2)
lsp(ggplot2)
# Search only for scale functions
lsp(ggplot2, pattern="scales")
}
}

